MOBILE-13: Training preparing screens, data analysis integration and debug training actualizing

See merge request 1837/android-app!20

+++ android/feature/trainings/airplaneGame/impl/src/main/java/ru/miem/psychoEvaluation/feature/trainings/airplaneGame/impl/game/entities/AirplaneView.kt
@@ -89,7 +89,7 @@ class AirplaneView(
         image.rotation(Angle.fromDegrees(airplaneRotationDegrees))
     }
 
-    fun onNewData(rawData: Int, speed: Double) {
+    fun onNewData(rawData: Double, speed: Double) {
         if (isAlive) {
             acceleration.y = speed - velocity.y
             velocity.y += speed * AIRPLANE_VELOCITY_MULTIPLIER
