+++ android/feature/trainings/airplaneGame/impl/src/main/java/ru/miem/psychoEvaluation/feature/trainings/airplaneGame/impl/game/entities/GameWorld.kt
@@ -72,7 +72,7 @@ class GameWorld(
         }
     }
 
-    fun onNewData(rawData: Int, speed: Double) = airplane.onNewData(rawData, speed)
+    fun onNewData(rawData: Double, speed: Double) = airplane.onNewData(rawData, speed)
 
     fun onDestroy() {
         airplane.onDestroy()
